requirements.txt
fastapi
uvicorn
python-dotenv==1.0.1
twilio==9.3.6
langchain==0.2.3
langchain-openai==0.1.8
pydantic==2.9.2
supabase==2.8.1
pandas==2.0.2